# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.FeedbackResponse:
    properties:
      admin_reply:
        type: string
      email:
        description: 联系邮箱
        type: string
      feedback_time:
        allOf:
        - $ref: '#/definitions/utils.JSONTime'
        description: 反馈时间
      form_content:
        description: 表单内容
      id:
        type: integer
      ip:
        type: string
      reply_time:
        allOf:
        - $ref: '#/definitions/utils.JSONTime'
        description: 回复时间
      report_type:
        type: string
      report_url:
        type: string
      status:
        type: string
      ticket_no:
        description: 工单号
        type: string
      user_agent:
        type: string
    type: object
...
  dto.MenuCreateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID，NULL表示主菜单
        type: integer
      sort:
        description: 排序，范围1-10
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址（主菜单可为空）
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
  dto.MenuDeleteRequest:
    properties:
      children_action:
        description: 子菜单处理方式：delete-删除子菜单，upgrade-升级为主菜单
        enum:
        - delete
        - upgrade
        type: string
    type: object
  dto.MenuResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  dto.MenuTreeNode:
    properties:
      children:
        description: 子菜单列表
        items:
          $ref: '#/definitions/dto.MenuTreeNode'
        type: array
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  dto.MenuUpdateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID
        type: integer
      sort:
        description: 排序
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
...
  dto.SubmitFeedbackRequest:
    properties:
      attachmentFiles:
        description: 附件文件URL数组
        items:
          type: string
        type: array
      description:
        type: string
      email:
        description: 可选联系邮箱
        type: string
      reason:
        description: 可选原因字段
        type: string
      reportType:
        enum:
        - copyright
        - inappropriate
        - summary
        - suggestion
        type: string
      reportUrl:
        type: string
    required:
    - description
    - reportType
    - reportUrl
    type: object
...
  dto.UpdateFeedbackRequest:
    properties:
      admin_reply:
        type: string
      status:
        enum:
        - pending
        - resolved
        - closed
        type: string
    required:
    - status
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  utils.JSONTime:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
...
  /admin/menus:
    get:
      consumes:
      - application/json
      description: 按类型获取菜单树
      parameters:
      - description: '菜单类型: aggregate/navigation/footer'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuTreeNode'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单树
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新的菜单项
      parameters:
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /admin/menus/{id}:
    delete:
      consumes:
      - application/json
      description: 硬删除菜单项，不可恢复
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 删除选项
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.MenuDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      description: 获取菜单项的详细信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 更新菜单项信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
...
  /api/v1/admin/feedback:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.FeedbackResponse'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取反馈列表
      tags:
      - 反馈管理
  /api/v1/admin/feedback/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除反馈
      tags:
      - 反馈管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FeedbackResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取反馈详情
      tags:
      - 反馈管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 反馈ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新反馈
      tags:
      - 反馈管理
...
  /api/v1/feedback:
    post:
      consumes:
      - application/json
      description: 用户提交反馈投诉
      parameters:
      - description: 反馈信息
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FeedbackResponse'
              type: object
      summary: 提交反馈投诉
      tags:
      - 反馈
  /api/v1/feedback/ticket/{ticket_no}:
    get:
      consumes:
      - application/json
      description: 根据工单号查询反馈投诉的处理进度和结果
      parameters:
      - description: 工单号
        in: path
        name: ticket_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FeedbackResponse'
              type: object
      summary: 查询反馈投诉进度
      tags:
      - 反馈
...
  /menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: '菜单类型: aggregate/navigation/footer (不传则返回所有类型)'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuTreeNode'
                  type: array
              type: object
      summary: 获取菜单树
      tags:
      - 菜单
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"